完成扫描器模块，优化前端体验

+++ api/resource.py
@@ -2,12 +2,13 @@
 __author__ = 'Joynice'
 from flask_restful import Resource
 from apps.cms.models import Task, User
-from models import Asset
 from . import field
 from .parse import HostServerPost_parse, HostServerDelete_parse, HostServerGet_parse, HostServerUpgrade_parse, \
     ResultGet_parse
 from exts import db
-from webscan.scan.cmstest import buildPayload
+from scan.webscan.Cms import web_scan
+from scan.hostscan.PortScan import host_scan
+from utils.web_tools import match_url
 
 class HostServer(Resource):
 
@@ -37,6 +38,7 @@ class HostServer(Resource):
                 return field.unauth_error(message='身份验证失败！')
         else:
             return field.params_error(message='身份验证参数缺失')
+
     def post(self):
         '''
         下发任务
@@ -44,6 +46,8 @@ class HostServer(Resource):
         '''
         args = HostServerPost_parse.parse_args()
         url = args.get('url')
+        if not match_url(url=url):
+            return field.params_error(message='URL格式不正确！')
         cycle = args.get('cycle')
         number = args.get('number')
         secret_key = args.get('secret_key')
@@ -55,7 +59,7 @@ class HostServer(Resource):
                 if user.is_api == 'ApiEnum.DOWN':
                     return field.unauth_error(message='该用户已经被禁用API，请联系超级管理员解决！')
                 if url and cycle and number:
-                    task = Task(url=url, cycle=str(cycle), number=number, user_id=user.id)
+                    task = Task(url=url, cycle=str(cycle), number=number, user_id=user.id, referer='API')
                     db.session.add(task)
                     try:
                         db.session.commit()
@@ -65,7 +69,8 @@ class HostServer(Resource):
                     result_id = task.result_id
                     if number == 1:
                         task.state = 'State.ING_SCAN'
-                        buildPayload.delay(url=url, taskid=task_id)
+                        web_scan.delay(url=url, taskid=task_id)
+                        host_scan.delay(url=url, taskid=task_id)
                     return field.success(message='下发任务成功，结果请稍后查询', data={'task_id': task_id, 'result_id': result_id})
                 else:
                     return field.params_error(message='参数缺失')
@@ -97,10 +102,17 @@ class HostServer(Resource):
                 if url or cycle or number:
                     task = Task.query.filter_by(task_id=task_id).first()
                     if task:
+                        if task.number == 1:
+                            return field.params_error('该任务正在扫描或者已经完成，无法更新任务！')
                         task.url = url or task.url
                         task.number = number or task.number
                         task.cycle = cycle or task.cycle
+                        task.referer = 'API'
                         db.session.commit()
+                        if task.number == 1:
+                            task.state = 'State.ING_SCAN'
+                            web_scan.delay(url=url, taskid=task_id)
+                            host_scan.delay(url=url, task_id=task_id)
                         return field.success(message='更新任务成功！')
                     else:
                         return field.params_error(message='没有找到该任务')
@@ -130,6 +142,8 @@ class HostServer(Resource):
                     return field.params_error(message='参数缺少')
                 task = Task.query.filter_by(task_id=task_id).first()
                 if task:
+                    if task.state == 'State.ING_SCAN':
+                        return field.params_error(message='任务正在进行中，无法删除！')
                     db.session.delete(task)
                     db.session.commit()
                     return field.success(message='删除任务成功')
@@ -149,6 +163,7 @@ class Result(Resource):
     def get(self):
         args = ResultGet_parse.parse_args()
         result_id = args.get('result_id')
+        print(result_id)
         secret_key = args.get('secret_key')
         if secret_key:
             user = User.query.filter_by(secret_key=secret_key).first()
@@ -159,11 +174,13 @@ class Result(Resource):
                     return field.unauth_error(message='该用户已经被禁用API，请联系超级管理员解决！')
                 if not result_id:
                     return field.params_error('参数缺失')
-                result1 = Task.query.filter_by(result_id=result_id).first()
-                if result1:
-                    result_data = result1.result
+                result = Task.query.filter_by(result_id=result_id).first()
+                if result:
+                    web_data = result.result
+                    cms_data = result.cms_result
+                    result1 = field.result_parse(cms_data, web_data)
                     return field.success(message='查询成功',
-                                         data={'task_id': result1.task_id, 'result_id': result_id, 'result': result_data})
+                                         data={'task_id': result.task_id, 'result_id': result_id, 'result': result1})
             else:
                 return field.unauth_error(message='身份验证失败！')
         else:
